Defined interface for item administration.
The purpose is to make item administration a replaceable piece of logic inside the adaptive quiz activity.
